+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Add signal handlers for SIGINT and SIGTERM to gracefully shut down the daemons
+- Add signal handlers to gracefully shut down the SIP and HTTP daemons
+- Move SIP daemon into its own thread using pthreads
 
 ### Tests
 - Clean up all valgrind warnings